Fix issue with example - index is out of range
At the moment, the indexer on line 34 is trying to get the 3rd element (second index) of the example multidimensional array. The issue is that the array only has 2 initial elements, therefore this code will throw an exception.  Changing this index from 2 to 1 fixes the issue, and the expected output on line 38 is now accurate